Services: Moved CourseSelectionStore
